Move import to personal settings for now
