+++ js/app.js
@@ -63,7 +63,6 @@
 		appBinds: function() {
 			// Add item bind
 			$('div.add-jot-item').on('click', function(e){ OCA.Jot.App.addItem(); });
-			$('button.jot-import').on('click', this._onClickImport);
 		},
 
 		/**
@@ -129,79 +128,6 @@
 		 */
 		updateLayout: function() {
 			this.container.isotope('layout');
-		},
-
-		/**
-		 * Handles clicking of the import button
-		 */
-		_onClickImport: function() {
-			// First show a dialog explaining the dealio
-			OC.dialogs.info(
-				'This wizard helps you import notes from Google Keep. </br>First, you need to visit Google Takeout to export your data. Click \'OK\' to visit the website. Once there, generate a .zip file containing only your Google Keep data, then upload this to your ownCloud.',
-				'Import Wizard',
-				function() {
-					// Take then to google takeout
-					window.open('https://takeout.google.com/settings/takeout?pli=1');
-					OC.dialogs.info(
-						'Next, upload this archive folder to your ownCloud in a new window and then return to this page. Click \'OK\' to open a new ownCloud window.',
-						'Import Wizard',
-						function() {
-							window.open(OC.generateUrl('/apps/files'));
-							OC.dialogs.info(
-								'Click \'OK\' and select the archive file you uploaded.',
-								'Import Wizard',
-								function() {
-									OC.dialogs.filepicker(
-										t('jot', 'Choose your exported zip file:'),
-										OCA.Jot.App.importFromZip,
-										false,
-										'application/zip'
-									);
-								},
-								false
-							);
-						},
-						false
-					);
-				},
-				false
-			);
-
-		},
-
-		/**
-		 * Handle triggering the import from a given path
-		 */
-		importFromZip: function(path) {
-			// Call ajax
-			var es = new OC.EventSource(OC.generateUrl('/apps/jot/api/1.0/jots/import'), {path: path});
-			// Fire a info dialog and keep the user updated with the progress
-			OC.dialogs.info(
-				'Loading...',
-				'Importing',
-				function() {
-					// Callback for what?, when you hit ok at the end, refresh the page
-					window.location.assign((OC.generateUrl('apps/jot')));
-				},
-				false
-			);
-
-			es.listen('progress', function(message) {
-				console.log(message);
-				// Some progress occured
-				$('#oc-dialog-3-content p').text(message);
-			});
-
-			es.listen('error', function(message) {
-				console.log(message);
-				$('#oc-dialog-3-content p').text('An error occured.');
-			});
-
-			es.listen('complete', function(message) {
-				$('#oc-dialog-3-content p').text('Complete! Click OK to reload the page.');
-			});
-
-			// Refresh the page to load the new shiny content
 		}
 
 	}
@@ -213,5 +139,4 @@
 // Go go go
 $(document).ready(function() {
 	OCA.Jot.App.initialize();
-
 });
