+++ settings.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace OCA\Jot\AppInfo;
+
+use OCP\Template;
+
+$page = new Template('jot', 'settings');
+return $page->fetchPage();
