+++ templates/settings.php
@@ -1,0 +1,10 @@
+<?php
+/** @var array $_ */
+script('jot', 'settings');
+?>
+
+<div class="section">
+    <a name="jot"></a>
+    <h2>Jot</h2>
+    <button class="jot-import">Import from Google Keep</button>
+</div>
